Add Makefile target for virtual env creation. (#37)
See #36 

This PR updates the Makefile with a new target for virtual env creation,
and includes the new target as dep of test target, so `make test` can be
run directly after checking out the code.

It's setup so that if either requirements file changes, it'll update the
virtual env, but otherwise it'll skip that step (based on timestamp
files).

It also updates the README so a new person knows how to run tests, both
via the Makefile target or via pytest manually.