Fix docker permissions for keycloak container. Fixes #39 (#40)
In #34 I fixed a logging issue so that we could see the keycloak logs,
but inadvertently broke the volume permissions, which was not noticed at
the time since I didn't re-create the volume from earlier.

This PR updates the volume to use a bind mount of the local data
directory, so it mounts at the expected point in the container (which is
owned by the `keycloak` user, so there are no longer permission issues
*and* the import file exists in the volume already (no need to copy it
in).

Tested with:

```
docker compose down -v
docker compose build --no-cache --pull
docker compose up -d
docker compose logs -f keycloak
```

(I haven't investigated how or why it passed on CI though - unless
volumes are being re-used there? Hope not).

See #39 for more info.