Add container publishing to Makefile and CI workflow. (#65)
This PR adds container publishing targets to the Makefile and uses those
from the CI workflow to publish the latest container image after a
successful run (on the main branch).

Note that currently it's just publishing the stuf-api and stuf-spa
images, not the custom keycloak image (I assume a project using stuf
should provide it's own identity provider - and can setup keycloak in
the same way if needed for dev.

As mentioned during review, the SPA image won't be much use currently, as we need to move the build-time config options to run-time (from the entrypoint).